requirements.txt
matplotlib==3.8.4
matplotlib-inline==0.1.7
numpy==1.26.4
yfinance==0.2.38
python-binance==1.0.19
python-dotenv==1.0.1
pandas==2.2.2
boto3==1.34.145
schedule==1.2.2